Further templatize the Makefile
If the author wants to change the configured build/ directory (or
sub-directories), or the src/ content directories, it should be easy to
do this. This commit adds additional variables to the top of the file,
which are then used in the main make logic.

This commit also removes the stylesheet variable, as that is a feature
that should come at a later date.